Final commit

+++ HomeWork1/circle.h
@@ -9,12 +9,12 @@
 ///
 /// @param radius   the radius of the circle
 /// @return         the area of the circle
- double getArea(int radius);
+ float  getArea(int radius);
 
 /// Get the circumference of a circle
 /// @param radius   the radius of the circle
 /// @return         the circumference of the circle
- double getCircumference(int radius);
+ float getCircumference(int radius);
 
 // CIRCLE_H
 #endif
